Refresh Control theme * Move dictionary merging code to NSAttributedString. * Update UITextField method to use new method. * Add theme support to UIRefreshControl. * Update README.d to include the two new theme options.
